% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{data_preparations}
\alias{data_preparations}
\title{Prepares dataframe with survey responses.}
\usage{
data_preparations(file)
}
\arguments{
\item{file}{String pointing to the csv file.}
}
\value{
Clean and tidy dataset with anonymised responses, ready to analise.
}
\description{
Reads a csv file containing qualtrics' export and cleans data.
}
