% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ratings.R
\name{get_image_ratings}
\alias{get_image_ratings}
\title{Creates a dataframe with images' ratings.}
\usage{
get_image_ratings(df)
}
\arguments{
\item{df}{Dataframe with survey's responses, after being cleaned by data_preparations().}
}
\value{
A long dataframe with image ratings and responses_id, ready to combine
  it with other sources.
}
\description{
Reads a clean dataframe with responses and generates a long
  dataset with all image ratings, associated to every response_id.
}
